# docs/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    description: Standard API response format used across all endpoints
    properties:
      code:
        description: '@Description Response code: 0 for success, 1 for error'
        example: 0
        type: integer
      jwt_token:
        description: '@Description JWT token when applicable'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      message:
        description: '@Description Response message'
        example: OK
        type: string
      response:
        description: '@Description Main response data (varies by endpoint)'
      summary:
        description: '@Description Summary data when applicable'
      sweatratesummary:
        description: '@Description Sweat rate summary when applicable'
      sweatsummary:
        description: '@Description Sweat summary when applicable'
      userdetails:
        description: '@Description User details when applicable'
      userid:
        description: '@Description User ID when applicable'
        example: 123
        type: integer
      users:
        description: '@Description Users list when applicable'
    type: object
  models.ClientHistoryRequest:
    properties:
      email:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  models.DetailedSummaryRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
...
  models.HistoryRequest:
    properties:
      email:
        type: string
      from_date:
        type: string
      to_date:
        type: string
      userid:
        type: integer
      username:
        type: string
    required:
    - email
    - from_date
    - to_date
    - userid
    - username
    type: object
...
  models.HydrationRequest:
    properties:
      device_type:
        type: integer
      email:
        type: string
      height:
        type: number
      image_id:
        type: integer
      image_path:
        type: string
      sweat_position:
        type: number
      time_taken:
        type: number
      userid:
        type: integer
      username:
        type: string
      weight:
        type: number
    required:
    - device_type
    - email
    - height
    - sweat_position
    - time_taken
    - userid
    - username
    - weight
    type: object
...
  models.HydrationSummaryRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
  models.HydrationSummaryResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      response:
        $ref: '#/definitions/models.HydrationSummaryData'
    type: object
...
  models.SummaryRequest:
    properties:
      email:
        type: string
      sweat_position:
        type: number
      username:
        type: string
    required:
    - email
    - sweat_position
    - username
    type: object
...
  models.UpdateHydrationRequest:
    properties:
      bmi:
        type: number
      device_type:
        type: integer
      email:
        type: string
      height:
        type: number
      id:
        type: integer
      sweat_loss:
        type: number
      sweat_position:
        type: number
      sweat_rate:
        type: number
      tbsa:
        type: number
      time_taken:
        type: number
      username:
        type: string
      weight:
        type: number
    required:
    - email
    - id
    - username
    type: object
...
  models.UpdateSweatDataRequest:
    properties:
      email:
        type: string
      image_id:
        type: integer
      sweat_loss:
        type: number
      sweat_rate:
        type: number
      userid:
        type: integer
      username:
        type: string
    required:
    - email
    - image_id
    - sweat_loss
    - sweat_rate
    - userid
    - username
    type: object
host: localhost:8500
info:
  contact:
    email: support@innovosens.com
    name: InnovoSens Support
    url: https://innovosens.com
  description: This is the InnovoSens REST API for hydration and sweat analysis data
    collection.
  license:
    name: InnovoSens License
    url: https://innovosens.com/license
  termsOfService: https://innovosens.com/terms
  title: InnovoSens API
  version: "1.0"
paths:
...
  /Services/protected/getClientHistory:
    post:
      consumes:
      - application/json
      description: Get recent history for a client
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Client history request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ClientHistoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get client history
      tags:
      - Reports
  /Services/protected/getElectrolyteHistory:
    post:
      consumes:
      - application/json
      description: Get electrolyte history for a date range (returns only creation_datetime
        and image_id)
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: History request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HistoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get electrolyte history
      tags:
      - Reports
  /Services/protected/getHyderartionHistory:
    post:
      consumes:
      - application/json
      description: Get hydration history for a date range with inclusive date handling
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: History request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HistoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get hydration history
      tags:
      - Reports
  /Services/protected/getHydrationSummaryScreen:
    post:
      consumes:
      - application/json
      description: Get formatted hydration data specifically for the summary screen
        display
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Summary screen request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HydrationSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HydrationSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get hydration summary screen data
      tags:
      - Hydration
  /Services/protected/getSummary:
    post:
      consumes:
      - application/json
      description: Get summary data based on sweat position
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Summary request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get summary
      tags:
      - Reports
...
  /Services/protected/getUserDetailedSummary:
    post:
      consumes:
      - application/json
      description: Get detailed summary with multiple data sources (hydration data,
        summary, sweat summary, sweat rate summary)
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Detailed summary request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DetailedSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get detailed summary
      tags:
      - Reports
  /Services/protected/innovoHyderation:
    post:
      consumes:
      - application/json
      description: Record basic hydration data with automatic BMI, TBSA, sweat loss,
        and sweat rate calculations
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Hydration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HydrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Record hydration data
      tags:
      - Hydration
  /Services/protected/newinnovoHyderation:
    post:
      consumes:
      - application/json
      description: Record enhanced hydration data with additional calculations and
        summaries
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Enhanced hydration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HydrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Record enhanced hydration data
      tags:
      - Hydration
  /Services/protected/updateHyderationValue:
    post:
      consumes:
      - application/json
      description: Update existing hydration data
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated hydration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateHydrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update hydration data
      tags:
      - Hydration
...
  /Services/protected/updateSweatData:
    post:
      consumes:
      - application/json
      description: Update sweat analysis data
      parameters:
      - description: Bearer JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sweat data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSweatDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update sweat data
      tags:
      - Hydration
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"